+++ lib/lexer.js
@@ -11,8 +11,8 @@ import {
   OPERATOR_TRIM,
   REF_RANGE
 } from './constants.js';
-import { lexers } from './lexerParts.js';
 import { mergeRefTokens } from './mergeRefTokens.js';
+import { lexers } from './lexers/sets.js';
 
 const isType = (t, type) => t && t.type === type;
 
@@ -132,71 +132,39 @@ export function getTokens (fx, tokenHandlers, options = {}) {
   const numHandlers = tokenHandlers.length;
   while (pos < fx.length) {
     const startPos = pos;
-    // const s = fx.slice(pos);
-    let tokenType = '';
-    let tokenValue = '';
+    let token;
     for (let i = 0; i < numHandlers; i++) {
-      const t = tokenHandlers[i](fx, pos, opts);
-      if (t) {
-        tokenType = t.type;
-        tokenValue = t.value;
-        pos += tokenValue.length;
+      token = tokenHandlers[i](fx, pos, opts);
+      if (token) {
+        pos += token.value.length;
         break;
       }
     }
 
-    if (!tokenType) {
-      tokenType = UNKNOWN;
-      tokenValue = fx[pos];
+    if (!token) {
+      token = {
+        type: UNKNOWN,
+        value: fx[pos]
+      };
       pos++;
     }
-
-    const token = withLocation
-      ? {
-        type: tokenType,
-        value: tokenValue,
-        loc: [ startPos, pos ]
-      }
-      : {
-        type: tokenType,
-        value: tokenValue
-      };
+    if (withLocation) {
+      token.loc = [ startPos, pos ];
+    }
 
     // make a note if we found a let/lambda call
-    if (lastToken && tokenValue === '(' && lastToken.type === FUNCTION) {
-      // const lastLC = lastToken.value.toLowerCase();
-      // if (lastLC === 'lambda' || lastLC === 'let') {
+    if (lastToken && token.value === '(' && lastToken.type === FUNCTION) {
       if (/^l(?:ambda|et)$/i.test(lastToken.value)) {
         letOrLambda++;
       }
     }
     // make a note if we found a R or C unknown
-    if (tokenType === UNKNOWN) {
-      const valLC = tokenValue.toLowerCase();
+    if (token.type === UNKNOWN && token.value.length === 1) {
+      const valLC = token.value.toLowerCase();
       unknownRC += (valLC === 'r' || valLC === 'c') ? 1 : 0;
     }
 
-    // check for termination
-    if (tokenType === STRING) {
-      const l = tokenValue.length;
-      if (tokenValue === '""') {
-        // common case that IS terminated
-      }
-      else if (tokenValue === '"' || tokenValue[l - 1] !== '"') {
-        token.unterminated = true;
-      }
-      else if (tokenValue !== '""' && tokenValue[l - 2] === '"') {
-        let p = l - 1;
-        while (tokenValue[p] === '"') { p--; }
-        const atStart = (p + 1);
-        const oddNum = ((l - p + 1) % 2 === 0);
-        if (!atStart !== oddNum) {
-          token.unterminated = true;
-        }
-      }
-    }
-
-    if (negativeNumbers && tokenType === NUMBER) {
+    if (negativeNumbers && token.type === NUMBER) {
       const last1 = lastToken;
       // do we have a number preceded by a minus?
       if (last1 && isType(last1, OPERATOR) && last1.value === '-') {
@@ -207,7 +175,7 @@ export function getTokens (fx, tokenHandlers, options = {}) {
           !causesBinaryMinus(tail1)
         ) {
           const minus = tokens.pop();
-          token.value = '-' + tokenValue;
+          token.value = '-' + token.value;
           if (token.loc) {
             // ensure offsets are up to date
             token.loc[0] = minus.loc[0];
