+++ lib/parseRef.js
@@ -9,7 +9,7 @@ import {
   REF_STRUCT,
   OPERATOR
 } from './constants.js';
-import { lexersRefs } from './lexerParts.js';
+import { lexersRefs } from './lexers/sets.js';
 import { getTokens } from './lexer.js';
 
 // Liberally split a context string up into parts.
