+++ lib/parseSRange.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-multi-spaces */
 /* eslint-disable no-undefined */
-import { WS } from './lexerParts.js';
+import { WS } from './lexers/lexWhitespace.js';
 
 const AT = 64; // @
 const BR_CLOSE = 93; // ]
